Initial attempt to get iLand work on Qt6
The bigger changes were:
- QRegExp: for now it is still there (compat5 layer), and partly switched to QRegularExpression
- for Q_PROPERTY for bite some more includes were necessary (incomplete types). in one case some Qt macros
- needed to update also the Windows KIT (not only the compiler)

it runs now for both GCC and MSVC (2022) with Qt 6.3.1

+++ src/bite/biteengine.cpp
@@ -86,7 +86,7 @@ void BiteEngine::setup()
         int lineno = result.property("lineNumber").toInt();
         QStringList code_lines = code.replace('\r', "").split('\n'); // remove CR, split by LF
         QString code_part;
-        for (int i=std::max(0, lineno - 5); i<std::min(lineno+5, code_lines.count()); ++i)
+        for (int i=std::max(0, lineno - 5); i<std::min(lineno+5, static_cast<int>(code_lines.count())); ++i)
             code_part.append(QString("%1: %2 %3\n").arg(i).arg(code_lines[i]).arg(i==lineno?"  <---- [ERROR]":""));
         qCCritical(biteSetup).noquote() << "Javascript Error in file" << result.property("fileName").toString() << ":" << result.property("lineNumber").toInt() << ":" << result.toString() << ":\n" << code_part;
         throw IException("BITE Error in Javascript (Please check the logfile): " + result.toString());
