+++ src/tools/globalsettings.h
@@ -163,7 +163,7 @@ const double cAutotrophicRespiration = 0.47;
 #define Globals (GlobalSettings::instance())
 
 // provide a hashing function for the QPoint type (needed from stand init functions, ABE, ...)
-inline uint qHash(const QPoint &key)
+inline size_t qHash(const QPoint &key)
 {
     return qHash(key.x()) ^ qHash(key.y());
 }
